rename dataclass folders
