+++ python/src/qdk_chemistry/algorithms/hamiltonian_unitary_builder/time_evolution/partially_randomized.py
@@ -27,7 +27,7 @@ import numpy as np
 
 from qdk_chemistry.algorithms.hamiltonian_unitary_builder.time_evolution.qdrift import QDrift
 from qdk_chemistry.data import QubitHamiltonian, Settings, UnitaryRepresentation
-from qdk_chemistry.data.hamiltonian_unitary.containers.pauli_product_formula import (
+from qdk_chemistry.data.unitary_representation.containers.pauli_product_formula import (
     ExponentiatedPauliTerm,
     PauliProductFormulaContainer,
 )
