+++ python/src/qdk_chemistry/algorithms/hamiltonian_unitary_builder/time_evolution/qdrift.py
@@ -20,7 +20,7 @@ import numpy as np
 
 from qdk_chemistry.algorithms.hamiltonian_unitary_builder.base import HamiltonianUnitaryBuilder, TimeEvolutionBuilder
 from qdk_chemistry.data import QubitHamiltonian, Settings, UnitaryRepresentation
-from qdk_chemistry.data.hamiltonian_unitary.containers.pauli_product_formula import (
+from qdk_chemistry.data.unitary_representation.containers.pauli_product_formula import (
     ExponentiatedPauliTerm,
     PauliProductFormulaContainer,
 )
