+++ python/src/qdk_chemistry/data/__init__.py
@@ -107,17 +107,17 @@ from qdk_chemistry._core.data import (
 from qdk_chemistry.data.base import DataClass
 from qdk_chemistry.data.circuit import Circuit
 from qdk_chemistry.data.circuit_executor_data import CircuitExecutorData
+from qdk_chemistry.data.controlled_unitary import ControlledUnitary
 from qdk_chemistry.data.encoding_validation import EncodingMismatchError, validate_encoding_compatibility
 from qdk_chemistry.data.enums.fermion_mode_order import FermionModeOrder
 from qdk_chemistry.data.estimator_data import EnergyExpectationResult, MeasurementData
-from qdk_chemistry.data.hamiltonian_unitary.base import UnitaryRepresentation
-from qdk_chemistry.data.hamiltonian_unitary.containers.base import UnitaryContainer
-from qdk_chemistry.data.hamiltonian_unitary.containers.pauli_product_formula import PauliProductFormulaContainer
-from qdk_chemistry.data.hamiltonian_unitary.controlled_unitary import ControlledUnitary
 from qdk_chemistry.data.noise_models import QuantumErrorProfile
 from qdk_chemistry.data.qpe_result import QpeResult
 from qdk_chemistry.data.qubit_hamiltonian import QubitHamiltonian
 from qdk_chemistry.data.symmetries import Symmetries
+from qdk_chemistry.data.unitary_representation.base import UnitaryRepresentation
+from qdk_chemistry.data.unitary_representation.containers.base import UnitaryContainer
+from qdk_chemistry.data.unitary_representation.containers.pauli_product_formula import PauliProductFormulaContainer
 
 # Give Users the option to use "Error" suffix for exceptions if they prefer
 SettingNotFoundError = SettingNotFound
