+++ python/src/qdk_chemistry/data/unitary_representation/containers/pauli_product_formula.py
