+++ python/tests/test_time_evolution_container.py
@@ -11,7 +11,7 @@ import h5py
 import numpy as np
 import pytest
 
-from qdk_chemistry.data.hamiltonian_unitary.containers.pauli_product_formula import (
+from qdk_chemistry.data.unitary_representation.containers.pauli_product_formula import (
     ExponentiatedPauliTerm,
     PauliProductFormulaContainer,
 )
